adjust task method

+++ src/routers/task.js
@@ -1,27 +1,49 @@
 const express = require('express');
 const Task = require('../models/task');
+const auth = require('../middleware/auth');
 const router = new express.Router();
 
-router.post('/tasks', (req, res) => {
-  const task = new Task(req.body);
-  task.save().then(() => res.send(task)).catch(e => res.status(500).send(e));
+router.post('/tasks', auth, async (req, res) => {
+  const task = new Task({
+    ...req.body,
+    owner: req.user._id
+  });
+
+  try {
+    await task.save();
+    res.status(201).send(task);
+  } catch (e) {
+    res.status(400).send(e);
+  }
 })
 
-router.get('/tasks', (req, res) => {
-  Task.find({}).then(tasks => res.send(tasks)).catch(e => res.status(500).send(e));
+router.get('/tasks', auth, async (req, res) => {
+  try {
+    // const tasks = await Task.find({ owner: req.user._id });
+    await req.user.populate('tasks').execPopulate();
+    res.send(req.user.tasks);
+  } catch (e) {
+    res.status(500).send();
+  }
 })
 
-router.get('/tasks/:id', (req, res) => {
+router.get('/tasks/:id', auth, async (req, res) => {
   const _id = req.params.id;
-  Task.findById(_id).then(task => {
+
+  try {
+    const task = await Task.findOne({ _id, owner: req.user._id });
+
     if (!task) {
       return res.status(404).send();
     }
+
     res.send(task);
-  }).catch(e => res.status(500).send())
+  } catch (e) {
+    res.status(500).send();
+  }
 })
 
-router.patch('/tasks/:id', async (req, res) => {
+router.patch('/tasks/:id', auth, async (req, res) => {
   const updates = Object.keys(req.body);
   const allowedUpdates = ['description', 'completed'];
   const isValidOperation = updates.every(update => allowedUpdates.includes(update));
@@ -31,20 +53,23 @@ router.patch('/tasks/:id', async (req, res) => {
   }
 
   try {
-    const task = await Task.findOneAndUpdate(req.params.id, req.body, { new: true, runValidators: true });
+    const task = await Task.findOne({ _id: req.params.id, owner: req.user._id });
+
     if (!task) {
       return res.status(404).send();
     }
 
+    updates.forEach(update => task[update] = req.body[update]);
+    await task.save();
     res.send(task);
   } catch (e) {
     res.status(400).send(e);
   }
 })
 
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id', auth, async (req, res) => {
   try {
-    const task = await Task.findByIdAndDelete(req.params.id);
+    const task = await Task.findByIdAndDelete({ _id: req.params.id, owner: req.user._id });
 
     if (!task) {
       return res.status(404).send();
